requirements.txt
django-5.0
celery-5.3.6
rabbitmq-server
user-agent
